+++ app/lib/theme/widgets/FileWidget.tsx
@@ -1,73 +1,13 @@
-import React, { MutableRefObject, useRef, useState } from 'react';
+import React, { useState } from 'react';
 import { useRouter } from 'next/router';
 import * as Sentry from '@sentry/nextjs';
 import { WidgetProps } from '@rjsf/core';
-import styled from 'styled-components';
-import { Button } from '@button-inc/bcgov-theme';
 import path from 'path';
 import { useCreateAttachment } from '../../../schema/mutations/attachment/createAttachment';
 import { useDeleteAttachment } from '../../../schema/mutations/attachment/deleteAttachment';
 
 import bytesToSize from '../../../utils/bytesToText';
-import { CancelIcon, LoadingSpinner } from '../../../components';
-
-const StyledContainer = styled('div')`
-  margin-top: 16px;
-  margin-bottom: 32px;
-  width: 100%;
-  display: flex;
-  justify-content: space-between;
-  border: 1px solid rgba(0, 0, 0, 0.16);
-  border-radius: 4px;
-  padding: 16px;
-`;
-
-const StyledDetails = styled('div')`
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-`;
-
-const StyledH4 = styled('h4')`
-  margin: 0;
-`;
-
-const StyledLink = styled.button`
-  color: ${(props) => props.theme.color.links};
-  text-decoration-line: underline;
-`;
-
-const StyledButton = styled(Button)`
-  min-width: 160px;
-  white-space: nowrap;
-  display: flex;
-  flex-direction: row;
-  justify-content: center;
-  margin-left: 8px;
-`;
-
-const StyledError = styled('div')`
-  color: #e71f1f;
-  margin-top: 10px;
-`;
-
-const StyledFileDiv = styled('div')`
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-
-  margin-left: 16px;
-  margin-top: 10px;
-  & svg {
-    margin: 0 8px;
-  }
-`;
-
-const StyledDeleteBtn = styled('button')`
-  &:hover {
-    opacity: 0.6;
-  }
-`;
+import FileComponent from '../components/FileComponent';
 
 type File = {
   id: string | number;
@@ -88,30 +28,6 @@ const checkFileType = (file, fileTypes) => {
   return typesArr.includes(extension);
 };
 
-const Error = ({ error, fileTypes }) => {
-  if (error === 'uploadFailed') {
-    return <StyledError>File failed to upload, please try again</StyledError>;
-  }
-
-  if (error === 'deleteFailed') {
-    return <StyledError>Delete file failed, please try again</StyledError>;
-  }
-
-  if (error === 'fileType') {
-    return (
-      <StyledError>
-        Please use an accepted file type. Accepted types for this field are:
-        <div>{fileTypes}</div>
-      </StyledError>
-    );
-  }
-
-  if (error === 'fileSize') {
-    return <StyledError>Files must be less than 100MB.</StyledError>;
-  }
-  return null;
-};
-
 const FileWidget: React.FC<FileWidgetProps> = ({
   id,
   disabled,
@@ -126,14 +42,13 @@ const FileWidget: React.FC<FileWidgetProps> = ({
   const [createAttachment, isCreatingAttachment] = useCreateAttachment();
   const [deleteAttachment, isDeletingAttachment] = useDeleteAttachment();
 
-  const hiddenFileInput = useRef() as MutableRefObject<HTMLInputElement>;
-  const allowMultipleFiles = uiSchema['ui:options']?.allowMultipleFiles;
-  const acceptedFileTypes = uiSchema['ui:options']?.fileTypes;
-  const buttonVariant = uiSchema['ui:options']?.buttonVariant || 'primary';
+  const allowMultipleFiles =
+    (uiSchema['ui:options']?.allowMultipleFiles as boolean) ?? false;
+  const acceptedFileTypes = (uiSchema['ui:options']?.fileTypes as string) ?? '';
+  const buttonVariant = (uiSchema['ui:options']?.buttonVariant ||
+    'primary') as string;
   const isFiles = value?.length > 0;
   const loading = isCreatingAttachment || isDeletingAttachment;
-  const isSecondary = buttonVariant === 'secondary';
-
   // 104857600 bytes = 100mb
   const maxFileSizeInBytes = 104857600;
 
@@ -265,24 +180,8 @@ const FileWidget: React.FC<FileWidgetProps> = ({
     e.target.value = '';
   };
 
-  const handleClick = () => {
-    hiddenFileInput.current.click();
-  };
-
   const showModal = () => {
     window.location.hash = 'file-error';
-  };
-  const buttonLabel = () => {
-    if (isFiles && !allowMultipleFiles) {
-      return 'Replace';
-    }
-    if (isFiles && allowMultipleFiles) {
-      return 'Add file';
-    }
-    if (allowMultipleFiles) {
-      return 'Upload(s)';
-    }
-    return 'Upload';
   };
 
   const handleDownload = async (uuid, fileName) => {
@@ -299,65 +198,22 @@ const FileWidget: React.FC<FileWidgetProps> = ({
   };
 
   return (
-    <StyledContainer
-      className="file-widget"
-      style={{ border: error && '1px solid #E71F1F' }}
-    >
-      <StyledDetails>
-        <StyledH4>{label}</StyledH4>
-        {isFiles &&
-          value.map((file: File) => (
-            <StyledFileDiv key={file.uuid}>
-              <StyledLink
-                data-testid="file-download-link"
-                onClick={(e) => {
-                  e.preventDefault();
-                  handleDownload(file.uuid, file.name);
-                }}
-              >
-                {file.name}
-              </StyledLink>
-              <StyledDeleteBtn
-                data-testid="file-delete-btn"
-                onClick={(e: React.MouseEvent<HTMLInputElement>) => {
-                  e.preventDefault();
-                  handleDelete(file.id);
-                }}
-                disabled={isDeletingAttachment}
-              >
-                <CancelIcon />
-              </StyledDeleteBtn>
-            </StyledFileDiv>
-          ))}
-        {error && <Error error={error} fileTypes={acceptedFileTypes} />}
-      </StyledDetails>
-      <div>
-        <StyledButton
-          id={`${id}-btn`}
-          onClick={(e: React.MouseEvent<HTMLInputElement>) => {
-            e.preventDefault();
-            handleClick();
-          }}
-          variant={buttonVariant}
-          disabled={isCreatingAttachment || disabled}
-        >
-          {loading ? (
-            <LoadingSpinner color={isSecondary ? '#000000' : '#fff'} />
-          ) : (
-            buttonLabel()
-          )}
-        </StyledButton>
-      </div>
-      <input
-        data-testid="file-test"
-        ref={hiddenFileInput}
-        onChange={handleChange}
-        style={{ display: 'none' }}
-        type="file"
-        required={required}
-        accept={acceptedFileTypes && acceptedFileTypes.toString()}
-      />
-    </StyledContainer>
+    <FileComponent
+      allowMultipleFiles={allowMultipleFiles as boolean}
+      isCreatingAttachment={isCreatingAttachment}
+      isDeletingAttachment={isDeletingAttachment}
+      error={error}
+      buttonVariant={buttonVariant}
+      handleDelete={handleDelete}
+      handleDownload={handleDownload}
+      onChange={handleChange}
+      disabled={disabled}
+      fileTypes={acceptedFileTypes}
+      id={id}
+      label={label}
+      required={required}
+      value={value}
+    />
   );
 };
